refactor: SessionManager.class의 getSession(HttpRequest) 삭제
