+++ src/main/java/servlet/view/support/ViewResolver.java
@@ -1,4 +1,4 @@
-package servlet.view;
+package servlet.view.support;
 
 import com.github.jknack.handlebars.Handlebars;
 import com.github.jknack.handlebars.Template;
@@ -6,6 +6,7 @@ import com.github.jknack.handlebars.io.ClassPathTemplateLoader;
 import com.github.jknack.handlebars.io.TemplateLoader;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import servlet.view.View;
 
 import java.io.IOException;
