+++ src/test/resources/HTTP_SESSION_TEST.txt
@@ -2,4 +2,4 @@ GET /user/list HTTP/1.1
 Host: localhost:8080
 Connection: keep-alive
 Accept: */*
-Cookie: logined=true; Path=/; SESSIONID=24f92f0e-eaad-4da6-9f68-cc019ecdfa37
+Cookie: logined=true; Path=/; SESSIONID=24f92f0e-eaad-4da6-9f68-cc019ecdfa37;
